+++ example/hello-rsp/webapp/helloView.rsp.dart
@@ -4,7 +4,8 @@ part of hello_rsp;
 
 /** Template, helloView, for rendering the view. */
 Future helloView(HttpConnect connect) { //#2
-  var _cs_ = new List<HttpConnect>(), request = connect.request, response = connect.response;
+  var _t0_, _cs_ = new List<HttpConnect>(),
+  request = connect.request, response = connect.response;
 
   if (!connect.isIncluded)
     response.headers.contentType = ContentType.parse("""text/html; charset=utf-8""");