+++ example/hello-templating/webapp/home.rsp.dart
@@ -4,7 +4,8 @@ part of hello_templating;
 
 /** Template, home, for rendering the view. */
 Future home(HttpConnect connect) { //#2
-  var _cs_ = new List<HttpConnect>(), request = connect.request, response = connect.response;
+  var _t0_, _cs_ = new List<HttpConnect>(),
+  request = connect.request, response = connect.response;
 
   if (!connect.isIncluded)
     response.headers.contentType = ContentType.parse("""text/html; charset=utf-8""");