+++ lib/stream.dart
@@ -8,6 +8,7 @@ import "dart:io";
 import "dart:async";
 import "dart:collection" show HashMap, LinkedHashMap;
 import "dart:uri";
+import "dart:json" as Json;
 import 'package:meta/meta.dart';
 import 'package:args/args.dart' show Options;
 import "package:logging/logging.dart" show Logger;