+++ test/syntax/include.rsp.dart
@@ -8,7 +8,8 @@ import 'package:stream/stream.dart';
 
 /** Template, include, for rendering the view. */
 Future include(HttpConnect connect, {foo, more, less}) { //#3
-  var _cs_ = new List<HttpConnect>(), request = connect.request, response = connect.response;
+  var _t0_, _cs_ = new List<HttpConnect>(),
+  request = connect.request, response = connect.response;
 
   if (!connect.isIncluded)
     response.headers.contentType = ContentType.parse("""text/html; charset=utf-8""");